pyproject.toml
[project]
name = "pid-parts"
version = "0.1.0"
description = "Convert P&ID PDFs to a deduplicated Markdown parts list (LangGraph POC)"
authors = [
    {name = "quantumleeps",email = "15917719+quantumleeps@users.noreply.github.com"}
]
license = "MIT"
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "langgraph (>=0.4.3,<0.5.0)",
    "pydantic (>=2.11.4,<3.0.0)",
    "pdf2image (>=1.17.0,<2.0.0)",
    "pillow (>=11.2.1,<12.0.0)",
    "numpy (>=2.2.5,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "langchain[openai] (>=0.3.25,<0.4.0)",
    "langchain-google-genai (>=2.1.4,<3.0.0)",
    "typer[all] (>=0.15.4,<0.16.0)",
    "pydantic-settings (>=2.9.1,<3.0.0)"
]

[tool.poetry]
packages = [{include = "pid_parts", from = "src"}]

[project.scripts]
pid-ingest = "pid_parts.cli:app"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
black = "^25.1.0"
ruff = "^0.11.9"
pre-commit = "^4.2.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
